serve s3: refactor to use build tags to exclude plan9 like /lib/kv
