+++ cmd/serve/s3/metastore_bolt_unsupported.go
@@ -1,0 +1,9 @@
+//go:build plan9 || js
+
+package s3
+
+import "fmt"
+
+func newBoltMetaStore(path string) (metadataStore, error) {
+	return nil, fmt.Errorf("persistent metadata database is not supported on this platform")
+}
