+++ cmd/serve/s3/metastore_test.go
@@ -1,7 +1,6 @@
 package s3
 
 import (
-	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -87,28 +86,4 @@ func testMetadataStore(t *testing.T, store metadataStore) {
 
 func TestMemoryMetaStore(t *testing.T) {
 	testMetadataStore(t, newMemoryMetaStore())
-}
-
-func TestBoltMetaStore(t *testing.T) {
-	dbPath := filepath.Join(t.TempDir(), "test-meta.db")
-	store, err := newBoltMetaStore(dbPath)
-	require.NoError(t, err)
-	testMetadataStore(t, store)
-}
-
-func TestBoltMetaStorePersistence(t *testing.T) {
-	dbPath := filepath.Join(t.TempDir(), "persist-meta.db")
-
-	store, err := newBoltMetaStore(dbPath)
-	require.NoError(t, err)
-	store.Store("mybucket/mykey", map[string]string{"Content-Type": "application/json"})
-	require.NoError(t, store.Close())
-
-	store2, err := newBoltMetaStore(dbPath)
-	require.NoError(t, err)
-	defer func() { _ = store2.Close() }()
-
-	got, ok := store2.Load("mybucket/mykey")
-	require.True(t, ok)
-	assert.Equal(t, "application/json", got["Content-Type"])
 }
